bug fixes
bug fixes

I am just exterminating some errors in the code
