+++ css/style.css
@@ -33,7 +33,7 @@ a{color: var(--dark);} a:hover{color: var(--ho);} a:foccus{color: var(--ho);} a:
   
 nav{  
     position: -webkit-sticky;
-    position: sticky;    
+    position: sticky;     
     background-color: var(--white);
     border: none;
     outline: none;
@@ -71,7 +71,7 @@ li{
 .bg{
     background: var(--ho); 
 }
-.btn{
+.btn{ 
   color: #fff; 
 }
 .sbtn{ 
@@ -97,7 +97,24 @@ li{
 .center{
   text-align: center; 
 }
- 
+.loader { 
+  position: fixed;
+  background-color: #996bfd;
+  display: none; 
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  height: 100vh;  
+  width: 100vw; 
+  z-index: 1200;   
+} 
+.loader h2{
+  position: absolute;
+  bottom: 10vh;
+  color: rgba(255, 255, 255, 0.966);
+  font-family: Arial, Helvetica, sans-serif;
+  font-size: 1.4rem;  
+}
 .form{
     width: 92%;
     max-width: 560px;
