+++ inc/Auth/index.html
@@ -7,9 +7,14 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>suplike Auth</title>
     <style>
+        body{
+            text-align: center;
+            text-indent: 2%; 
+        }
         .logo {
             background: url("../../img/logo.png"); 
             background-repeat: no-repeat;
+            background-size: cover; 
             border-radius: 50%;
             width: 160px;
             height: 160px;
