+++ sql/suplike.sql
@@ -18,10 +18,10 @@ SET time_zone = "+00:00";
 /*!40101 SET NAMES utf8mb4 */;
 
 --
--- Database: `bethro_suplike`
+-- Database: `suplike` 
 --
-CREATE DATABASE IF NOT EXISTS `bethro_suplike` DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci;
-USE `bethro_suplike`;
+CREATE DATABASE IF NOT EXISTS `suplike` DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci;
+USE `suplike`;
 
 -- --------------------------------------------------------
